add gins processing node in launch file
