+++ .gitignore
@@ -1,0 +1,5 @@
+build/
+devel/
+log/
+*.bag
+traj/
